requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
openpyxl==3.1.2
gspread==5.12.0
google-auth==2.25.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
pydantic[email]==2.5.3
pandas==2.1.4
